+++ GyroData/Component/GraphView/GraphView.swift
@@ -7,43 +7,6 @@
 
 import UIKit
 
-
-//그래프의 포인트 데이터를 저장,사용하기 위한 클래스
-class GraphBuffer {
-    
-    var array: [CGFloat?]
-    var index = 0
-    var count: Int {
-        return array.count
-    }
-    //입력한 count 크기 만큼 배열을 생성한다
-    init(count: Int) {
-        array = Array(repeating: nil, count: count)
-    }
-    
-    //배열 초기화 메소드
-    func resetToValue(_ value: CGFloat?) {
-        let count = array.count
-        array = Array(repeating: value, count: count)
-    }
-    //현재 인덱스에 element값을 넣는다
-    //index는 호출될때 마다 1씩 증가하며, count 만큼 증가한다면 다시 index는 0을 향하게 된다.
-    func write(_ element: CGFloat) {
-        array[index % array.count] = element
-        index += 1
-    }
-    //현재 그래프의 경로 값들을 넘겨준다
-    //현재 인덱스의 값부터 시작되는 새로운 배열을 만들고, 고차 함수를 이용해 옵셔널바인딩 하여 리턴 하게 된다
-    //이전 경로의 마지막 값이 index 0 이 된다
-    func nextItems() -> [CGFloat] {
-        var result = Array<CGFloat?>()
-        for loop in 0..<array.count {
-            result.append(array[(loop+index) % array.count])
-        }
-        return result.compactMap { $0 }
-    }
-}
-
 class GraphView: UIView {
     
     public var maxValue: CGFloat = 20
@@ -77,43 +40,39 @@ class GraphView: UIView {
     
     required init?(coder: NSCoder) {
         super.init(coder: coder)
+        doInitSetup()
     }
     
-    //그래프 속성 설정
-    func doInitSetup() {
-        designLayer(xLayer, color: .red)
-        designLayer(yLayer, color: .green)
-        designLayer(zLayer, color: .blue)
-    }
-    
-    private func designLayer(_ layer: CAShapeLayer, color: UIColor) {
-        layer.strokeColor = color.cgColor
-        layer.fillColor = UIColor.clear.cgColor
-        layer.lineWidth = 2
-        layer.lineCap = .round
-    }
-    
-    //그래프 초기화 기능 메소드
-    public func reset() {
+    // MARK: - open func
+    func reset() {
         guard let layer = self.layer as? CAShapeLayer else { return }
         graphSwipeAnimation = false
         aPoint?.resetToValue(nil)
         bPoint?.resetToValue(nil)
         cPoint?.resetToValue(nil)
         layer.path = makePath().aPath
+        layer.path = makePath().bPath
+        layer.path = makePath().cPath
+        
+    }
+        
+    func justShowGraph() {
+        graphData.forEach { data in
+            let x = CGFloat(data.coodinate.x)
+            let y = CGFloat(data.coodinate.y)
+            let z = CGFloat(data.coodinate.z)
+            addNewValue(aValue: x, bValue: y, cValue: z)
+        }
     }
     
     //이전 좌표, 현재 좌표를 이용하여 변화를 애니메이션 처리
-    public func animateNewValue(aValue: CGFloat, bValue: CGFloat, cValue: CGFloat, duration: Double = 0.0) {
+    func addNewValue(aValue: CGFloat, bValue: CGFloat, cValue: CGFloat, duration: Double = 0.0) {
         let animation = CABasicAnimation(keyPath: "strokeEnd")
         
         //value가 최대 혹은 최소 값을 초과 했을 경우 스케일 재설정
         resetScale([aValue,bValue,cValue])
-        
         let oldPathInfo = makePath()
-        
         animation.duration = duration
-        
         // 이전 경로로 부터 애니메이션 시작
         animation.fromValue = oldPathInfo.aPath
         
@@ -131,7 +90,6 @@ class GraphView: UIView {
             yLayer.add(animation, forKey: animation.keyPath)
             zLayer.add(animation, forKey: animation.keyPath)
         }
- 
         
         // 새 좌표 업데이트
         aPoint?.write(aValue)
@@ -147,31 +105,34 @@ class GraphView: UIView {
         }
     }
     
-    public func showGraph() {
-        
-        graphData.forEach { data in
-            let x = CGFloat(data.coodinate.x)
-            let y = CGFloat(data.coodinate.y)
-            let z = CGFloat(data.coodinate.z)
-            animateNewValue(aValue: x, bValue: y, cValue: z)
-        }
     
+    // MARK: - private func
+    private func doInitSetup() {
+        designLayer(xLayer, color: .red)
+        designLayer(yLayer, color: .green)
+        designLayer(zLayer, color: .blue)
     }
     
-    //경로 값을 이용해 선을 그린다
+    private func designLayer(_ layer: CAShapeLayer, color: UIColor) {
+        layer.strokeColor = color.cgColor
+        layer.fillColor = UIColor.clear.cgColor
+        layer.lineWidth = 2
+        layer.lineCap = .round
+    }
+    
+    
     private func makePath() -> (aPath: CGPath, bPath: CGPath, cPath: CGPath, xInterval: CGFloat) {
         guard let aPoint = aPoint, let bPoint = bPoint, let cPoint = cPoint else {
             (layer as? CAShapeLayer)?.path = nil
             return (UIBezierPath().cgPath, UIBezierPath().cgPath, UIBezierPath().cgPath, 0)
         }
-//
         let xInterval = bounds.width / (CGFloat(aPoint.count) - 1)
         let range = minValue - maxValue
         let yInterval = bounds.height / range
- 
+        
         for (idx, value) in aPoint.nextItems().enumerated() {
-
-            let x = xForIndex(idx, xInterval) //x는 같고 y가 여러개여야한다.
+            
+            let x = xForIndex(idx, xInterval)
             if !graphSwipeAnimation && bounds.width <= x {
                 graphSwipeAnimation = true
             }
@@ -196,7 +157,7 @@ class GraphView: UIView {
                 cPath.move(to: CGPoint(x: x, y: cY))
             }
         }
-
+        
         self.layer.addSublayer(xLayer)
         self.layer.addSublayer(yLayer)
         self.layer.addSublayer(zLayer)
@@ -204,16 +165,15 @@ class GraphView: UIView {
         return (aPath.cgPath, bPath.cgPath, cPath.cgPath, xInterval)
     }
     
-    func xForIndex(_ index: Int, _ xInterval: CGFloat) -> CGFloat {
+    private func xForIndex(_ index: Int, _ xInterval: CGFloat) -> CGFloat {
         return CGFloat(index) * xInterval + bounds.origin.x
     }
     
-    func yForValue(_ value: CGFloat, _ yInterval: CGFloat) -> CGFloat {
+    private func yForValue(_ value: CGFloat, _ yInterval: CGFloat) -> CGFloat {
         return bounds.height / 2 + value * yInterval + bounds.origin.y
     }
     
-    func resetScale(_ motionData: [Double]) {
-
+    private func resetScale(_ motionData: [Double]) {
         for value in motionData {
             if value > maxValue {
                 maxValue = value + (value * 0.2)
@@ -224,6 +184,6 @@ class GraphView: UIView {
                 minValue = value + (value * 0.2)
             }
         }
-
+        
     }
 }
